# BanckEnd_Provider_managment-main/swagger.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/agreement/{id}": {
            "get": {
                "operationId": "AgreementController_getAgreementById",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AgreementModel"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "agreement"
                ]
            }
        },
...
        "/offer/provider/{id}": {
            "post": {
                "operationId": "OfferController_create",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateOfferDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "offer"
                ]
            },
            "get": {
                "operationId": "OfferController_getOfferByProviderId",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "offer"
                ]
            }
        },
        "/offer/agreement/{id}": {
            "get": {
                "operationId": "OfferController_getAllOfferByAgreementId",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "offer"
                ]
            }
        },
        "/offer/{id}": {
            "get": {
                "operationId": "OfferController_getOfferById",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "offer"
                ]
            }
        },
        "/offer/update/{id}": {
            "put": {
                "operationId": "OfferController_updateOfferById",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateOfferDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OfferEntity"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "offer"
                ]
            }
        },
        "/materialGroup": {
            "post": {
                "operationId": "MaterialGroupController_create",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateMaterialGroupDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MaterialGroupModel"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "materialGroup"
                ]
            },
            "get": {
                "operationId": "MaterialGroupController_getAll",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "materialGroup"
                ]
            }
        },
        "/materialGroup/update/{id}": {
            "put": {
                "operationId": "MaterialGroupController_updateOfferById",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateMaterialGroupDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MaterialGroupEntity"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "materialGroup"
                ]
            }
        },
        "/deal/offer/{id}": {
            "post": {
                "operationId": "DealController_create",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DealModel"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "deal"
                ]
            }
        },
...
        "/provider/{id}": {
            "get": {
                "operationId": "AuthProviderController_get",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProviderEntity"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "provider"
                ]
            }
        },
...
    },
    "info": {
        "title": "Provider-management-platform",
        "description": "API to handle everything related to provider-management",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "bearer": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http"
            }
        },
        "schemas": {
...
            "AgreementModel": {
                "type": "object",
                "properties": {}
            },
...
            "CreateOfferDto": {
                "type": "object",
                "properties": {
                    "agreementId": {
                        "type": "number"
                    },
                    "skills": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "experienceLevel": {
                        "type": "string"
                    },
                    "salary": {
                        "type": "string"
                    },
                    "category": {
                        "type": "string"
                    },
                    "rule": {
                        "type": "string"
                    }
                },
                "required": [
                    "agreementId",
                    "skills",
                    "experienceLevel",
                    "salary",
                    "category",
                    "rule"
                ]
            },
            "AgreementEntity": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number"
                    },
                    "title": {
                        "type": "string"
                    },
                    "position": {
                        "type": "string"
                    },
                    "providerName": {
                        "type": "string"
                    },
                    "providerEmail": {
                        "type": "string"
                    },
                    "skill": {
                        "type": "string"
                    },
                    "salary": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "cycle": {
                        "type": "string"
                    },
                    "materialGroup": {
                        "type": "string"
                    },
                    "teamMember": {
                        "type": "string"
                    },
                    "technologyLevel": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "jobStartDate": {
                        "type": "string"
                    },
                    "jobEndDate": {
                        "type": "string"
                    },
                    "startContractDate": {
                        "type": "string"
                    },
                    "review": {
                        "type": "number"
                    },
                    "comment": {
                        "type": "string"
                    },
                    "endContractDate": {
                        "type": "string"
                    },
                    "createdAt": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "updatedAt": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "deal": {
                        "$ref": "#/components/schemas/DealEntity"
                    }
                },
                "required": [
                    "id",
                    "title",
                    "position",
                    "providerName",
                    "providerEmail",
                    "skill",
                    "salary",
                    "description",
                    "cycle",
                    "materialGroup",
                    "teamMember",
                    "technologyLevel",
                    "role",
                    "status",
                    "jobStartDate",
                    "jobEndDate",
                    "startContractDate",
                    "endContractDate",
                    "createdAt",
                    "updatedAt",
                    "deal"
                ]
            },
            "DealEntity": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number"
                    },
                    "provider": {
                        "$ref": "#/components/schemas/ProviderEntity"
                    },
                    "agreement": {
                        "$ref": "#/components/schemas/AgreementEntity"
                    },
                    "offer": {
                        "$ref": "#/components/schemas/OfferEntity"
                    },
                    "review": {
                        "type": "number"
                    },
                    "createdAt": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "updatedAt": {
                        "format": "date-time",
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "provider",
                    "agreement",
                    "offer",
                    "review",
                    "createdAt",
                    "updatedAt"
                ]
            },
            "OfferEntity": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number"
                    },
                    "agreementId": {
                        "type": "number"
                    },
                    "skills": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "experienceLevel": {
                        "type": "string"
                    },
                    "salary": {
                        "type": "string"
                    },
                    "category": {
                        "type": "string"
                    },
                    "rule": {
                        "type": "string"
                    },
                    "createdAt": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "updatedAt": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "provider": {
                        "$ref": "#/components/schemas/ProviderEntity"
                    },
                    "deal": {
                        "$ref": "#/components/schemas/DealEntity"
                    }
                },
                "required": [
                    "id",
                    "agreementId",
                    "skills",
                    "experienceLevel",
                    "salary",
                    "category",
                    "rule",
                    "createdAt",
                    "updatedAt",
                    "provider",
                    "deal"
                ]
            },
            "ProviderEntity": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number"
                    },
                    "name": {
                        "type": "string"
                    },
                    "address": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "rule": {
                        "type": "string"
                    },
                    "contactName": {
                        "type": "string"
                    },
                    "role": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "existSince": {
                        "type": "string"
                    },
                    "validFrom": {
                        "type": "string"
                    },
                    "validUntil": {
                        "type": "string"
                    },
                    "masterAgreementType": {
                        "type": "string"
                    },
                    "review": {
                        "type": "number"
                    },
                    "createdAt": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "updatedAt": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "offers": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OfferEntity"
                        }
                    },
                    "deal": {
                        "$ref": "#/components/schemas/DealEntity"
                    }
                },
                "required": [
                    "id",
                    "name",
                    "address",
                    "phone",
                    "rule",
                    "contactName",
                    "role",
                    "email",
                    "existSince",
                    "validFrom",
                    "validUntil",
                    "masterAgreementType",
                    "review",
                    "createdAt",
                    "updatedAt",
                    "offers",
                    "deal"
                ]
            },
            "CreateMaterialGroupDto": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "price": {
                        "type": "string"
                    }
                },
                "required": [
                    "name",
                    "description",
                    "price"
                ]
            },
            "MaterialGroupModel": {
                "type": "object",
                "properties": {}
            },
            "MaterialGroupEntity": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "price": {
                        "type": "string"
                    },
                    "createdAt": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "updatedAt": {
                        "format": "date-time",
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "name",
                    "description",
                    "price",
                    "createdAt",
                    "updatedAt"
                ]
            },
            "DealModel": {
                "type": "object",
                "properties": {}
            },
...
        }
    }
}